feat: reusable share pda modal
