+++ src/app/(light)/dashboard/components/share-pda-modal/share-pda-form-error.tsx
@@ -7,7 +7,7 @@ type Props = {
   error: string;
 };
 
-export default function ShareCopyFormError({ error }: Props) {
+export default function SharePdaFormError({ error }: Props) {
   return (
     <>
       <Stack>
